chore: improve query performance
