+++ pressbooks-multi-institution.php
@@ -21,19 +21,19 @@ use PressbooksMultiInstitution\Database\Migration;
 use PressbooksMultiInstitution\Models\InstitutionUser;
 
 // TODO: Check if this is the best way to check for Pressbooks.
-//if (!class_exists('PressbooksMultiInstitution\Bootstrap')) {
-if (file_exists(__DIR__ . '/vendor/autoload.php')) {
-    require_once __DIR__ . '/vendor/autoload.php';
-} else {
-    $title = __('Missing dependencies', 'pressbooks-multi-institution');
-    $body = __(
-        'Please run <code>composer install</code> from the root of the plugin directory.',
-        'pressbooks-multi-institution'
-    );
+if (!class_exists('PressbooksMultiInstitution\Bootstrap')) {
+    if (file_exists(__DIR__ . '/vendor/autoload.php')) {
+        require_once __DIR__ . '/vendor/autoload.php';
+    } else {
+        $title = __('Missing dependencies', 'pressbooks-multi-institution');
+        $body = __(
+            'Please run <code>composer install</code> from the root of the plugin directory.',
+            'pressbooks-multi-institution'
+        );
 
-    wp_die("<h1>{$title}</h1><p>{$body}</p>");
+        wp_die("<h1>{$title}</h1><p>{$body}</p>");
+    }
 }
-//}
 
 register_activation_hook(__FILE__, [Migration::class, 'migrate']);
 register_activation_hook(__FILE__, function () {
